Move Size class to own file in interface layer
