+++ src/Main/InterfaceLayer/MappersToDomainElements/SizeInPixelsMapper.ts
@@ -1,0 +1,24 @@
+import { ISizeInPixels } from "../../HtmlMapper/ParagraphElements/Html_Image/Html_Image_SizeUnits"
+import { DocsInlineObjectSizeModel } from "../../Models/DocsDocumentModel"
+
+export class SizeInPixels implements ISizeInPixels{
+        height:number
+        width:number
+        unit:"pixel" = "pixel"
+    
+        static of(width:number, height:number):SizeInPixels{
+            return new SizeInPixels(width, height)
+        }
+        static convertDocsSizeToPixelSize(size:DocsInlineObjectSizeModel):SizeInPixels{
+            const widthInPoints = size?.width?.magnitude
+            const heightInPoints = size?.height?.magnitude
+            return SizeInPixels.of( SizeInPixels.convertPointSizetoPixelSize(widthInPoints), SizeInPixels.convertPointSizetoPixelSize(heightInPoints) )
+        }
+        private static convertPointSizetoPixelSize(pointSize:number):number{
+            return pointSize * (4/3)
+        }
+        constructor(width:number, height:number){
+            this.height = height
+            this.width = width
+        }
+    }
